More moving towards using signed types.
I'm still feeling this out, but I am starting to like the general idea.

+++ AK/Bitmap.h
@@ -10,12 +10,12 @@ namespace AK {
 class Bitmap {
 public:
     // NOTE: A wrapping Bitmap won't try to free the wrapped data.
-    static Bitmap wrap(byte* data, unsigned size)
+    static Bitmap wrap(byte* data, int size)
     {
         return Bitmap(data, size);
     }
 
-    static Bitmap create(unsigned size, bool default_value = 0)
+    static Bitmap create(int size, bool default_value = 0)
     {
         return Bitmap(size, default_value);
     }
@@ -27,13 +27,13 @@ public:
         m_data = nullptr;
     }
 
-    unsigned size() const { return m_size; }
-    bool get(unsigned index) const
+    int size() const { return m_size; }
+    bool get(int index) const
     {
         ASSERT(index < m_size);
         return 0 != (m_data[index / 8] & (1u << (index % 8)));
     }
-    void set(unsigned index, bool value) const
+    void set(int index, bool value) const
     {
         ASSERT(index < m_size);
         if (value)
@@ -46,17 +46,17 @@ public:
     const byte* data() const { return m_data; }
 
 private:
-    explicit Bitmap(unsigned size, bool default_value)
+    explicit Bitmap(int size, bool default_value)
         : m_size(size)
         , m_owned(true)
     {
         ASSERT(m_size != 0);
-        size_t size_to_allocate = ceil_div(size, 8u);
+        int size_to_allocate = ceil_div(size, 8);
         m_data = reinterpret_cast<byte*>(kmalloc(size_to_allocate));
         memset(m_data, default_value ? 0xff : 0x00, size_to_allocate);
     }
 
-    Bitmap(byte* data, unsigned size)
+    Bitmap(byte* data, int size)
         : m_data(data)
         , m_size(size)
         , m_owned(false)
@@ -64,7 +64,7 @@ private:
     }
 
     byte* m_data { nullptr };
-    unsigned m_size { 0 };
+    int m_size { 0 };
     bool m_owned { false };
 };
